second migration, for new column purchases
make sure that knex is installed globally so you can use knex as a command in your terminal. Then knex migrate:make add_   and the name of the column file. this creates a new migration file which you need to edit, as in this commit, then npm run migrate:latest once or twice and you will have a new column in your table of data